Split combine_intervals into own rs file and cleanup older implementations
